performed renamings
