+++ push/sender/src/main/java/org/jboss/aerogear/unifiedpush/message/TokenLoader.java
@@ -4,6 +4,7 @@ import java.util.List;
 import java.util.Set;
 import java.util.TreeSet;
 
+import javax.ejb.Stateless;
 import javax.enterprise.event.Event;
 import javax.enterprise.event.Observes;
 import javax.inject.Inject;
@@ -11,11 +12,14 @@ import javax.inject.Inject;
 import org.jboss.aerogear.unifiedpush.api.Variant;
 import org.jboss.aerogear.unifiedpush.dao.BatchException;
 import org.jboss.aerogear.unifiedpush.dao.ResultsStream;
+import org.jboss.aerogear.unifiedpush.message.holder.MessageHolderWithTokens;
+import org.jboss.aerogear.unifiedpush.message.holder.MessageHolderWithVariants;
 import org.jboss.aerogear.unifiedpush.message.jms.Dequeue;
 import org.jboss.aerogear.unifiedpush.message.jms.DispatchToQueue;
 import org.jboss.aerogear.unifiedpush.service.ClientInstallationService;
 import org.jboss.aerogear.unifiedpush.utils.AeroGearLogger;
 
+@Stateless
 public class TokenLoader {
 
     private final static int NUMBER_OF_BATCHES = 10;
@@ -28,13 +32,13 @@ public class TokenLoader {
 
     @Inject
     @DispatchToQueue
-    private Event<MessageWithTokens> dispatchTokensEvent;
+    private Event<MessageHolderWithTokens> dispatchTokensEvent;
 
     @Inject
     @DispatchToQueue
-    private Event<MessageWithVariants> nextBatchEvent;
+    private Event<MessageHolderWithVariants> nextBatchEvent;
 
-    public void loadAndQueueTokenBatch(@Observes @Dequeue MessageWithVariants msg) {
+    public void loadAndQueueTokenBatch(@Observes @Dequeue MessageHolderWithVariants msg) {
         final UnifiedPushMessage message = msg.getUnifiedPushMessage();
         final List<Variant> variants = msg.getVariants();
         final String lastTokenFromPreviousBatch = msg.getLastTokenFromPreviousBatch();
@@ -61,11 +65,11 @@ public class TokenLoader {
                         tokens.add(lastTokenInBatch);
                         tokensLoaded += 1;
                     }
-                    dispatchTokensEvent.fire(new MessageWithTokens(msg.getPushMessageInformation(), message, variant, tokens));
+                    dispatchTokensEvent.fire(new MessageHolderWithTokens(msg.getPushMessageInformation(), message, variant, tokens));
                 }
                 // should we load next batch ?
                 if (tokensLoaded >= NUMBER_OF_BATCHES * BATCH_SIZE) {
-                    nextBatchEvent.fire(new MessageWithVariants(msg.getPushMessageInformation(), message, msg.getVariantType(), variants, lastTokenInBatch));
+                    nextBatchEvent.fire(new MessageHolderWithVariants(msg.getPushMessageInformation(), message, msg.getVariantType(), variants, lastTokenInBatch));
                 }
             } catch (BatchException e) {
                 logger.severe("Failed to load batch of tokens", e);
