+++ push/sender/src/main/java/org/jboss/aerogear/unifiedpush/message/holder/MessageHolderWithTokens.java
@@ -1,36 +1,26 @@
-package org.jboss.aerogear.unifiedpush.message;
+package org.jboss.aerogear.unifiedpush.message.holder;
 
 import java.io.Serializable;
 import java.util.Collection;
 
 import org.jboss.aerogear.unifiedpush.api.PushMessageInformation;
 import org.jboss.aerogear.unifiedpush.api.Variant;
+import org.jboss.aerogear.unifiedpush.message.UnifiedPushMessage;
 
-public class MessageWithTokens implements Serializable {
+public class MessageHolderWithTokens extends AbstractMessageHolder implements Serializable {
 
     private static final long serialVersionUID = -7955411139315335655L;
 
-    private PushMessageInformation pushMessageInformation;
-    private UnifiedPushMessage unifiedPushMessage;
     private Variant variant;
     private Collection<String> deviceTokens;
 
-    public MessageWithTokens(PushMessageInformation pushMessageInformation, UnifiedPushMessage unifiedPushMessage, Variant variant, Collection<String> deviceTokens) {
+    public MessageHolderWithTokens(PushMessageInformation pushMessageInformation, UnifiedPushMessage unifiedPushMessage, Variant variant, Collection<String> deviceTokens) {
+        super(pushMessageInformation, unifiedPushMessage);
         if (!(deviceTokens instanceof Serializable)) {
             throw new IllegalArgumentException("deviceTokens must be a serializable collection");
         }
-        this.pushMessageInformation = pushMessageInformation;
-        this.unifiedPushMessage = unifiedPushMessage;
         this.variant = variant;
         this.deviceTokens = deviceTokens;
-    }
-
-    public PushMessageInformation getPushMessageInformation() {
-        return pushMessageInformation;
-    }
-
-    public UnifiedPushMessage getUnifiedPushMessage() {
-        return unifiedPushMessage;
     }
 
     public Variant getVariant() {
