+++ push/sender/src/main/java/org/jboss/aerogear/unifiedpush/message/holder/MessageHolderWithVariants.java
@@ -1,4 +1,4 @@
-package org.jboss.aerogear.unifiedpush.message;
+package org.jboss.aerogear.unifiedpush.message.holder;
 
 import java.io.Serializable;
 import java.util.ArrayList;
@@ -7,40 +7,29 @@ import java.util.Collection;
 import org.jboss.aerogear.unifiedpush.api.PushMessageInformation;
 import org.jboss.aerogear.unifiedpush.api.Variant;
 import org.jboss.aerogear.unifiedpush.api.VariantType;
+import org.jboss.aerogear.unifiedpush.message.UnifiedPushMessage;
 
-public class MessageWithVariants implements Serializable {
+public class MessageHolderWithVariants extends AbstractMessageHolder implements Serializable {
 
     private static final long serialVersionUID = -7955411139315335655L;
 
-    private PushMessageInformation pushMessageInformation;
-    private UnifiedPushMessage unifiedPushMessage;
     private VariantType variantType;
     private ArrayList<Variant> variants;
     private String lastTokenFromPreviousBatch;
 
-    public MessageWithVariants(PushMessageInformation pushMessageInformation, UnifiedPushMessage unifiedPushMessage, VariantType variantType, Collection<Variant> variants) {
+    public MessageHolderWithVariants(PushMessageInformation pushMessageInformation, UnifiedPushMessage unifiedPushMessage, VariantType variantType, Collection<Variant> variants) {
         this(pushMessageInformation, unifiedPushMessage, variantType, variants, null);
     }
 
-
-    public MessageWithVariants(PushMessageInformation pushMessageInformation, UnifiedPushMessage unifiedPushMessage, VariantType variantType, Collection<Variant> variants, String lastTokenFromPreviousBatch) {
-        this.pushMessageInformation = pushMessageInformation;
-        this.unifiedPushMessage = unifiedPushMessage;
+    public MessageHolderWithVariants(PushMessageInformation pushMessageInformation, UnifiedPushMessage unifiedPushMessage, VariantType variantType, Collection<Variant> variants, String lastTokenFromPreviousBatch) {
+        super(pushMessageInformation, unifiedPushMessage);
         this.variantType = variantType;
         this.variants = new ArrayList<Variant>(variants);
         this.lastTokenFromPreviousBatch = lastTokenFromPreviousBatch;
     }
 
-    public PushMessageInformation getPushMessageInformation() {
-        return pushMessageInformation;
-    }
-
     public VariantType getVariantType() {
         return variantType;
-    }
-
-    public UnifiedPushMessage getUnifiedPushMessage() {
-        return unifiedPushMessage;
     }
 
     public ArrayList<Variant> getVariants() {
