+++ push/sender/src/main/java/org/jboss/aerogear/unifiedpush/message/jms/MessageHolderWithTokensProducer.java
@@ -12,39 +12,39 @@ import javax.jms.Session;
 
 import org.jboss.aerogear.unifiedpush.api.VariantType;
 import org.jboss.aerogear.unifiedpush.message.MessageDeliveryException;
-import org.jboss.aerogear.unifiedpush.message.MessageWithVariants;
+import org.jboss.aerogear.unifiedpush.message.holder.MessageHolderWithTokens;
 
 @Stateless
-public class MessageWithVariantsDispatcher {
+public class MessageHolderWithTokensProducer {
 
     @Resource(mappedName = "java:/ConnectionFactory")
     private ConnectionFactory connectionFactory;
 
-    @Resource(mappedName = "java:/queue/AdmPushMessageQueue")
-    private Queue admPushMessageQueue;
+    @Resource(mappedName = "java:/queue/AdmTokenBatchQueue")
+    private Queue admTokenBatchQueue;
 
-    @Resource(mappedName = "java:/queue/APNsPushMessageQueue")
-    private Queue apnsPushMessageQueue;
+    @Resource(mappedName = "java:/queue/APNsTokenBatchQueue")
+    private Queue apnsTokenBatchQueue;
 
-    @Resource(mappedName = "java:/queue/GCMPushMessageQueue")
-    private Queue gcmPushMessageQueue;
+    @Resource(mappedName = "java:/queue/GCMTokenBatchQueue")
+    private Queue gcmTokenBatchQueue;
 
-    @Resource(mappedName = "java:/queue/MPNSPushMessageQueue")
-    private Queue mpnsPushMessageQueue;
+    @Resource(mappedName = "java:/queue/MPNSTokenBatchQueue")
+    private Queue mpnsTokenBatchQueue;
 
-    @Resource(mappedName = "java:/queue/SimplePushMessageQueue")
-    private Queue simplePushMessageQueue;
+    @Resource(mappedName = "java:/queue/SimplePushTokenBatchQueue")
+    private Queue simplePushTokenBatchQueue;
 
-    @Resource(mappedName = "java:/queue/WNSPushMessageQueue")
-    private Queue wnsPushMessageQueue;
+    @Resource(mappedName = "java:/queue/WNSTokenBatchQueue")
+    private Queue wnsTokenBatchQueue;
 
-    public void queueMessageVariantForProcessing(@Observes @DispatchToQueue MessageWithVariants msg) {
+    public void queueMessageVariantForProcessing(@Observes @DispatchToQueue MessageHolderWithTokens msg) {
         Connection connection = null;
         try {
             connection = connectionFactory.createConnection();
             Session session = connection.createSession(false, Session.AUTO_ACKNOWLEDGE);
-            Queue pushMessageQueue = selectQueue(msg.getVariantType());
-            MessageProducer messageProducer = session.createProducer(pushMessageQueue);
+            Queue tokenBatchQueue = selectQueue(msg.getVariant().getType());
+            MessageProducer messageProducer = session.createProducer(tokenBatchQueue);
             connection.start();
             messageProducer.send(session.createObjectMessage(msg));
         } catch (JMSException e) {
@@ -63,17 +63,17 @@ public class MessageWithVariantsDispatcher {
     private Queue selectQueue(VariantType variantType) {
         switch (variantType) {
             case ADM:
-                return admPushMessageQueue;
+                return admTokenBatchQueue;
             case ANDROID:
-                return gcmPushMessageQueue;
+                return gcmTokenBatchQueue;
             case IOS:
-                return apnsPushMessageQueue;
+                return apnsTokenBatchQueue;
             case SIMPLE_PUSH:
-                return simplePushMessageQueue;
+                return simplePushTokenBatchQueue;
             case WINDOWS_MPNS:
-                return mpnsPushMessageQueue;
+                return mpnsTokenBatchQueue;
             case WINDOWS_WNS:
-                return wnsPushMessageQueue;
+                return wnsTokenBatchQueue;
             default:
                 throw new IllegalStateException("Unknown variant type queue");
         }
