+++ push/sender/src/main/java/org/jboss/aerogear/unifiedpush/message/jms/MessageHolderWithVariantsConsumer.java
@@ -10,24 +10,24 @@ import javax.jms.MessageListener;
 import javax.jms.ObjectMessage;
 
 import org.jboss.aerogear.unifiedpush.message.MessageDeliveryException;
-import org.jboss.aerogear.unifiedpush.message.MessageWithTokens;
 import org.jboss.aerogear.unifiedpush.message.TokenLoader;
+import org.jboss.aerogear.unifiedpush.message.holder.MessageHolderWithVariants;
 import org.jboss.aerogear.unifiedpush.utils.AeroGearLogger;
 
-public class MessageWithTokensMDB implements MessageListener {
+public class MessageHolderWithVariantsConsumer implements MessageListener {
 
     private final AeroGearLogger logger = AeroGearLogger.getInstance(TokenLoader.class);
 
     @Inject
     @Dequeue
-    private Event<MessageWithTokens> message;
+    private Event<MessageHolderWithVariants> message;
 
     @Override
-    @TransactionAttribute(TransactionAttributeType.NOT_SUPPORTED)
+    @TransactionAttribute(TransactionAttributeType.REQUIRED)
     public void onMessage(Message jmsMessage) {
         try {
-            if (jmsMessage instanceof ObjectMessage && ((ObjectMessage) jmsMessage).getObject() instanceof MessageWithTokens) {
-                message.fire((MessageWithTokens) ((ObjectMessage) jmsMessage).getObject());
+            if (jmsMessage instanceof ObjectMessage && ((ObjectMessage) jmsMessage).getObject() instanceof MessageHolderWithVariants) {
+                message.fire((MessageHolderWithVariants) ((ObjectMessage) jmsMessage).getObject());
             } else {
                 logger.warning("Received message of wrong type: " + jmsMessage.getClass().getName());
             }
