+++ push/sender/src/main/java/org/jboss/aerogear/unifiedpush/message/jms/MessageHolderWithVariantsProducer.java
@@ -12,39 +12,39 @@ import javax.jms.Session;
 
 import org.jboss.aerogear.unifiedpush.api.VariantType;
 import org.jboss.aerogear.unifiedpush.message.MessageDeliveryException;
-import org.jboss.aerogear.unifiedpush.message.MessageWithTokens;
+import org.jboss.aerogear.unifiedpush.message.holder.MessageHolderWithVariants;
 
 @Stateless
-public class MessageWithTokensDispatcher {
+public class MessageHolderWithVariantsProducer {
 
     @Resource(mappedName = "java:/ConnectionFactory")
     private ConnectionFactory connectionFactory;
 
-    @Resource(mappedName = "java:/queue/AdmTokenBatchQueue")
-    private Queue admTokenBatchQueue;
+    @Resource(mappedName = "java:/queue/AdmPushMessageQueue")
+    private Queue admPushMessageQueue;
 
-    @Resource(mappedName = "java:/queue/APNsTokenBatchQueue")
-    private Queue apnsTokenBatchQueue;
+    @Resource(mappedName = "java:/queue/APNsPushMessageQueue")
+    private Queue apnsPushMessageQueue;
 
-    @Resource(mappedName = "java:/queue/GCMTokenBatchQueue")
-    private Queue gcmTokenBatchQueue;
+    @Resource(mappedName = "java:/queue/GCMPushMessageQueue")
+    private Queue gcmPushMessageQueue;
 
-    @Resource(mappedName = "java:/queue/MPNSTokenBatchQueue")
-    private Queue mpnsTokenBatchQueue;
+    @Resource(mappedName = "java:/queue/MPNSPushMessageQueue")
+    private Queue mpnsPushMessageQueue;
 
-    @Resource(mappedName = "java:/queue/SimplePushTokenBatchQueue")
-    private Queue simplePushTokenBatchQueue;
+    @Resource(mappedName = "java:/queue/SimplePushMessageQueue")
+    private Queue simplePushMessageQueue;
 
-    @Resource(mappedName = "java:/queue/WNSTokenBatchQueue")
-    private Queue wnsTokenBatchQueue;
+    @Resource(mappedName = "java:/queue/WNSPushMessageQueue")
+    private Queue wnsPushMessageQueue;
 
-    public void queueMessageVariantForProcessing(@Observes @DispatchToQueue MessageWithTokens msg) {
+    public void queueMessageVariantForProcessing(@Observes @DispatchToQueue MessageHolderWithVariants msg) {
         Connection connection = null;
         try {
             connection = connectionFactory.createConnection();
             Session session = connection.createSession(false, Session.AUTO_ACKNOWLEDGE);
-            Queue tokenBatchQueue = selectQueue(msg.getVariant().getType());
-            MessageProducer messageProducer = session.createProducer(tokenBatchQueue);
+            Queue pushMessageQueue = selectQueue(msg.getVariantType());
+            MessageProducer messageProducer = session.createProducer(pushMessageQueue);
             connection.start();
             messageProducer.send(session.createObjectMessage(msg));
         } catch (JMSException e) {
@@ -63,17 +63,17 @@ public class MessageWithTokensDispatcher {
     private Queue selectQueue(VariantType variantType) {
         switch (variantType) {
             case ADM:
-                return admTokenBatchQueue;
+                return admPushMessageQueue;
             case ANDROID:
-                return gcmTokenBatchQueue;
+                return gcmPushMessageQueue;
             case IOS:
-                return apnsTokenBatchQueue;
+                return apnsPushMessageQueue;
             case SIMPLE_PUSH:
-                return simplePushTokenBatchQueue;
+                return simplePushMessageQueue;
             case WINDOWS_MPNS:
-                return mpnsTokenBatchQueue;
+                return mpnsPushMessageQueue;
             case WINDOWS_WNS:
-                return wnsTokenBatchQueue;
+                return wnsPushMessageQueue;
             default:
                 throw new IllegalStateException("Unknown variant type queue");
         }
