+++ servers/ups-wildfly/src/main/webapp/WEB-INF/jboss-ejb3.xml
@@ -9,8 +9,8 @@
     <enterprise-beans>
         <!-- Push Message Queue MDBs -->
         <message-driven>
-            <ejb-name>AdmPushMessageMDB</ejb-name>
-            <ejb-class>org.jboss.aerogear.unifiedpush.message.jms.MessageWithVariantsMDB</ejb-class>
+            <ejb-name>AdmPushMessageConsumer</ejb-name>
+            <ejb-class>org.jboss.aerogear.unifiedpush.message.jms.MessageHolderWithVariantsConsumer</ejb-class>
             <transaction-type>Container</transaction-type>
             <activation-config>
                 <activation-config-property>
@@ -28,8 +28,8 @@
             </activation-config>
         </message-driven>
         <message-driven>
-            <ejb-name>APNsPushMessageMDB</ejb-name>
-            <ejb-class>org.jboss.aerogear.unifiedpush.message.jms.MessageWithVariantsMDB</ejb-class>
+            <ejb-name>APNsPushMessageConsumer</ejb-name>
+            <ejb-class>org.jboss.aerogear.unifiedpush.message.jms.MessageHolderWithVariantsConsumer</ejb-class>
             <transaction-type>Container</transaction-type>
             <activation-config>
                 <activation-config-property>
@@ -47,8 +47,8 @@
             </activation-config>
         </message-driven>
         <message-driven>
-            <ejb-name>GCMPushMessageMDB</ejb-name>
-            <ejb-class>org.jboss.aerogear.unifiedpush.message.jms.MessageWithVariantsMDB</ejb-class>
+            <ejb-name>GCMPushMessageConsumer</ejb-name>
+            <ejb-class>org.jboss.aerogear.unifiedpush.message.jms.MessageHolderWithVariantsConsumer</ejb-class>
             <transaction-type>Container</transaction-type>
             <activation-config>
                 <activation-config-property>
@@ -66,8 +66,8 @@
             </activation-config>
         </message-driven>
         <message-driven>
-            <ejb-name>MPNSPushMessageMDB</ejb-name>
-            <ejb-class>org.jboss.aerogear.unifiedpush.message.jms.MessageWithVariantsMDB</ejb-class>
+            <ejb-name>MPNSPushMessageConsumer</ejb-name>
+            <ejb-class>org.jboss.aerogear.unifiedpush.message.jms.MessageHolderWithVariantsConsumer</ejb-class>
             <transaction-type>Container</transaction-type>
             <activation-config>
                 <activation-config-property>
@@ -85,8 +85,8 @@
             </activation-config>
         </message-driven>
         <message-driven>
-            <ejb-name>SimplePushMessageMDB</ejb-name>
-            <ejb-class>org.jboss.aerogear.unifiedpush.message.jms.MessageWithVariantsMDB</ejb-class>
+            <ejb-name>SimplePushMessageConsumer</ejb-name>
+            <ejb-class>org.jboss.aerogear.unifiedpush.message.jms.MessageHolderWithVariantsConsumer</ejb-class>
             <transaction-type>Container</transaction-type>
             <activation-config>
                 <activation-config-property>
@@ -104,8 +104,8 @@
             </activation-config>
         </message-driven>
         <message-driven>
-            <ejb-name>WNSPushMessageMDB</ejb-name>
-            <ejb-class>org.jboss.aerogear.unifiedpush.message.jms.MessageWithVariantsMDB</ejb-class>
+            <ejb-name>WNSPushMessageConsumer</ejb-name>
+            <ejb-class>org.jboss.aerogear.unifiedpush.message.jms.MessageHolderWithVariantsConsumer</ejb-class>
             <transaction-type>Container</transaction-type>
             <activation-config>
                 <activation-config-property>
@@ -125,8 +125,8 @@
         
         <!-- Token Batch Queue MDBs -->
         <message-driven>
-            <ejb-name>AdmTokenBatchMDB</ejb-name>
-            <ejb-class>org.jboss.aerogear.unifiedpush.message.jms.MessageWithTokensMDB</ejb-class>
+            <ejb-name>AdmTokenBatchConsumer</ejb-name>
+            <ejb-class>org.jboss.aerogear.unifiedpush.message.jms.MessageHolderWithTokensConsumer</ejb-class>
             <transaction-type>Container</transaction-type>
             <activation-config>
                 <activation-config-property>
@@ -144,8 +144,8 @@
             </activation-config>
         </message-driven>
         <message-driven>
-            <ejb-name>APNsTokenBatchMDB</ejb-name>
-            <ejb-class>org.jboss.aerogear.unifiedpush.message.jms.MessageWithTokensMDB</ejb-class>
+            <ejb-name>APNsTokenBatchConsumer</ejb-name>
+            <ejb-class>org.jboss.aerogear.unifiedpush.message.jms.MessageHolderWithTokensConsumer</ejb-class>
             <transaction-type>Container</transaction-type>
             <activation-config>
                 <activation-config-property>
@@ -163,8 +163,8 @@
             </activation-config>
         </message-driven>
         <message-driven>
-            <ejb-name>GCMTokenBatchMDB</ejb-name>
-            <ejb-class>org.jboss.aerogear.unifiedpush.message.jms.MessageWithTokensMDB</ejb-class>
+            <ejb-name>GCMTokenBatchConsumer</ejb-name>
+            <ejb-class>org.jboss.aerogear.unifiedpush.message.jms.MessageHolderWithTokensConsumer</ejb-class>
             <transaction-type>Container</transaction-type>
             <activation-config>
                 <activation-config-property>
@@ -182,8 +182,8 @@
             </activation-config>
         </message-driven>
         <message-driven>
-            <ejb-name>MPNSTokenBatchMDB</ejb-name>
-            <ejb-class>org.jboss.aerogear.unifiedpush.message.jms.MessageWithTokensMDB</ejb-class>
+            <ejb-name>MPNSTokenBatchConsumer</ejb-name>
+            <ejb-class>org.jboss.aerogear.unifiedpush.message.jms.MessageHolderWithTokensConsumer</ejb-class>
             <transaction-type>Container</transaction-type>
             <activation-config>
                 <activation-config-property>
@@ -201,8 +201,8 @@
             </activation-config>
         </message-driven>
         <message-driven>
-            <ejb-name>SimplePushTokenBatchMDB</ejb-name>
-            <ejb-class>org.jboss.aerogear.unifiedpush.message.jms.MessageWithTokensMDB</ejb-class>
+            <ejb-name>SimplePushTokenBatchConsumer</ejb-name>
+            <ejb-class>org.jboss.aerogear.unifiedpush.message.jms.MessageHolderWithTokensConsumer</ejb-class>
             <transaction-type>Container</transaction-type>
             <activation-config>
                 <activation-config-property>
@@ -220,8 +220,8 @@
             </activation-config>
         </message-driven>
         <message-driven>
-            <ejb-name>WNSTokenBatchMDB</ejb-name>
-            <ejb-class>org.jboss.aerogear.unifiedpush.message.jms.MessageWithTokensMDB</ejb-class>
+            <ejb-name>WNSTokenBatchConsumer</ejb-name>
+            <ejb-class>org.jboss.aerogear.unifiedpush.message.jms.MessageHolderWithTokensConsumer</ejb-class>
             <transaction-type>Container</transaction-type>
             <activation-config>
                 <activation-config-property>
